Finished task, task list item and note tests

+++ app/controllers/api/v1/note_controller.rb
@@ -1,31 +1,43 @@
 class Api::V1::NoteController < ApplicationController
   def create
-    @note = Note.new(environment_id: params[:environment_id])
+    if user_in_environment?(params[:environment_id])
+      @note = Note.new(environment_id: params[:environment_id])
 
-    if @note.save
-      render_response("success", "OK!")
+      if @note.save
+        render_response("success", "OK!")
+      else
+        render_response("error", "Failed while creating a note")
+      end
     else
-      render_response("error", "Failed while creating a task list")
+      render_response("error", "User isn't in the environment!")
     end
   end
 
   def update
-    @note = Note.find(params[:id])
+    if user_in_environment?(params[:environment_id])
+      @note = Note.find(params[:id])
 
-    if @note.update(note_params)
-      render_response("success", "OK!")
+      if @note.update(note_params)
+        render_response("success", "OK!")
+      else
+        render_response("error", "Try again!")
+      end
     else
-      render_response("error", "Try again!")
+      render_response("error", "User isn't in the environment!")
     end
   end
 
   def destroy
-    @note = Note.find(params[:id])
+    if user_in_environment?(params[:environment_id])
+      @note = Note.find(params[:id])
     
-    if @note.destroy
-      render_response("success", "OK!")
+      if @note.destroy
+        render_response("success", "OK!")
+      else
+        render_response("error", "Try again!")
+      end
     else
-      render_response("error", "Try again!")
+      render_response("error", "User isn't in the environment!")
     end
   end
